Extract common parts of insight search into fragment macros
